+++ src/main/java/com/javajober/fileBlock/controller/FileBlockController.java
@@ -4,8 +4,8 @@ import com.javajober.core.error.exception.Exception404;
 import com.javajober.core.message.ErrorMessage;
 import com.javajober.core.message.SuccessMessage;
 import com.javajober.core.util.ApiUtils;
-import com.javajober.fileBlock.dto.request.FileBlockSaveRequests;
-import com.javajober.fileBlock.dto.request.FileBlockUpdateRequests;
+import com.javajober.fileBlock.filedto.FileBlockSaveRequests;
+import com.javajober.fileBlock.filedto.FileBlockUpdateRequests;
 import com.javajober.fileBlock.dto.response.FileBlockResponse;
 import com.javajober.fileBlock.dto.response.FileBlockResponses;
 import com.javajober.fileBlock.service.FileBlockService;
