+++ src/main/java/com/javajober/fileBlock/filedto/FileBlockSaveRequest.java
@@ -1,19 +1,18 @@
-package com.javajober.fileBlock.dto.request;
+package com.javajober.fileBlock.filedto;
 
 import com.javajober.fileBlock.domain.FileBlock;
 import lombok.Getter;
 
 @Getter
-public class FileBlockUpdateRequest {
+public class FileBlockSaveRequest {
 
-    private Long fileBlockId;
     private String fileTitle;
     private String fileDescription;
 
-    public FileBlockUpdateRequest() {
+    public FileBlockSaveRequest() {
     }
 
-    public static FileBlock toEntity(FileBlockUpdateRequest fileBlockSaveRequest, String fileName) {
+    public static FileBlock toEntity(FileBlockSaveRequest fileBlockSaveRequest, String fileName) {
         return FileBlock.builder()
                 .fileTitle(fileBlockSaveRequest.getFileTitle())
                 .fileDescription(fileBlockSaveRequest.getFileDescription())
