+++ src/main/java/com/javajober/fileBlock/filedto/FileBlockUpdateRequest.java
@@ -1,18 +1,19 @@
-package com.javajober.fileBlock.dto.request;
+package com.javajober.fileBlock.filedto;
 
 import com.javajober.fileBlock.domain.FileBlock;
 import lombok.Getter;
 
 @Getter
-public class FileBlockSaveRequest {
+public class FileBlockUpdateRequest {
 
+    private Long fileBlockId;
     private String fileTitle;
     private String fileDescription;
 
-    public FileBlockSaveRequest() {
+    public FileBlockUpdateRequest() {
     }
 
-    public static FileBlock toEntity(FileBlockSaveRequest fileBlockSaveRequest, String fileName) {
+    public static FileBlock toEntity(FileBlockUpdateRequest fileBlockSaveRequest, String fileName) {
         return FileBlock.builder()
                 .fileTitle(fileBlockSaveRequest.getFileTitle())
                 .fileDescription(fileBlockSaveRequest.getFileDescription())
