+++ src/main/java/com/javajober/fileBlock/filedto/FileBlockUpdateRequests.java
@@ -1,4 +1,4 @@
-package com.javajober.fileBlock.dto.request;
+package com.javajober.fileBlock.filedto;
 
 import lombok.Getter;
