+++ src/main/java/com/javajober/fileBlock/service/FileBlockService.java
@@ -4,10 +4,10 @@ import com.javajober.core.config.FileDirectoryConfig;
 import com.javajober.core.error.exception.Exception500;
 import com.javajober.core.message.ErrorMessage;
 import com.javajober.fileBlock.domain.FileBlock;
-import com.javajober.fileBlock.dto.request.FileBlockSaveRequest;
-import com.javajober.fileBlock.dto.request.FileBlockSaveRequests;
-import com.javajober.fileBlock.dto.request.FileBlockUpdateRequest;
-import com.javajober.fileBlock.dto.request.FileBlockUpdateRequests;
+import com.javajober.fileBlock.filedto.FileBlockSaveRequest;
+import com.javajober.fileBlock.filedto.FileBlockSaveRequests;
+import com.javajober.fileBlock.filedto.FileBlockUpdateRequest;
+import com.javajober.fileBlock.filedto.FileBlockUpdateRequests;
 import com.javajober.fileBlock.dto.response.FileBlockResponse;
 import com.javajober.fileBlock.dto.response.FileBlockResponses;
 import com.javajober.fileBlock.repository.FileBlockRepository;
